Reconfigure docker-compose-http.yml to allow for https testing
For Box.com development we need to be able to have a local https://
url to redirect to. This new configurate runs the Django development
server in debug mode (i.e., with reload) and nginx as an http proxy
(as opposed to a wsgi proxy as in qa/production.) Nginx runs on 80
and 443, using a self-signed certificate.