complete rewrite of tasks
rewrote init tasks to combine into single task because of need to relate metadata. BoxSyncAgent is rewritten to utilize the Box User Events Stream instead of relying explicitly on Box Webhooks. we still use Box Webhooks to trigger syncing the Events Stream, but the Webhooks are not the data source we are acting on anymore.