Added email confirmation field.
Added a field for users to confirm their email address, like we do for passwords.  Added a check to make sure that email and confirmEmail are the same before allowing user to register account.

Added a task to send an email to a settings.NEW_ACCOUNT_ALERT_EMAIL.  This email includes username, full name, email address, and ID.

Added an environment variable NEW_ACCOUNT_ALERT_EMAIL that defaults to the no-reply address.  This is the email that will receive email alerts when new accounts are created.  Make sure to set this before deploying.  The new_user_alert celery task will be called when a new user logs in and doesnt have an Agave token object.  This makes sure that it is the first time that a user is logging in.