Replace "start date" filters with simple show/hide history toggle. (#930)
Users were confused by the default view and could not easily figure out
how to see what is happening right now in the facilities.  They were being
presented by default with the full history of all projects in all facilities,
sorted from newest to oldest, causing projects scheduled a year or more into
the future to be the first thing displayed when they load the page.

This change removes the "start date"-based filters entirely with a simple
checkbox to show or hide projects that are already completed.  It also
reverses the default ordering from "newest first" to "oldest first".

When combined with applying the "hide history" filter by default, users
are now presented with an initial view that shows them as the first thing
they see on the page the projects that are in facilities right now and
future projects that haven't started yet the further down the page they go.

Co-authored-by: Jake Rosenberg <jrosenberg@tacc.utexas.edu>